Auto Aiming
Added Auto aiming based on vision calculations. This is a basic calculation not taking into account latency compensation or robot state estimation. That is a future endeavor. Put in a table of values which works extremely well on practice field at home -- very accurate! Also added a new amp scoring command which, go figure, doesn't really work. Also added the auto aiming command to the auto so if it's off a little bit it should be able to correct for that. We can shoot from anywhere the april tags are visible now!