all nighter with cyclone part 2: electric boogaloo
so. much. auto. tuning.

auto 1 + a new state estimation auto sorta work with state estimation. honestly have to test after the robot is back in working order. made a new command to update state estimation in auto. also made a soft limit override for the pitch for the shot right in front of the speaker. finished implementing the lob shot preset.